add scale and mask field into protobuf and bigdl module
